Added summary comments to DomainService class and methods
Added summary comments to the `DomainService` class and its methods in the `DomainService.cs` file. The class comment explains that it provides services for managing domain membership of the machine. The `JoinToDomain` method comment explains that it joins the machine to a domain, taking a `DomainJoinRequest` as a parameter. The `UnjoinFromDomain` method comment explains that it unjoins the machine from a domain, taking a `DomainUnjoinRequest` as a parameter.