Sort computers lists in Home class
In the `Home` class, the lists of available and unavailable computers (`_availableComputers` and `_unavailableComputers`) are now sorted by the `Name` property of the `computer` objects they contain. This is achieved by adding two new lines of code. After these changes, an asynchronous call to `StateHasChanged` method is made to notify the framework that the state has changed and the UI should be updated.