Update log message in `ConnectAs` method
Updated the log message that is displayed when calling the `ConnectAs` method with the `Intention.ReceiveThumbnail` parameter. The previous message, "Calling ConnectAs with Intention.GetThumbnail for {IPAddress}", has been changed to "Calling ConnectAs with Intention.ReceiveThumbnail for {IPAddress}". This change improves the accuracy of the log message to reflect the action being performed.