"Updated computer object handling in Home class"
Added code to handle computer object based on certain conditions in the `Home` class. When a specific condition is not met, the `computer.Thumbnail` is set to `null` and the `computer` object is moved from the `_availableComputers` list to the `_unavailableComputers` list. In a different scenario, when another condition is not met, the `computer.Thumbnail` is set to `null` and the `computer` object is added to the `tempUnavailable` list.