Added process start feature to Task Manager
Added a new feature to the Task Manager interface in the `TaskManager.razor` file that allows users to enter a process path and start a process. A new private string `_processPath` has been added to the `TaskManager.razor.cs` file to hold the user's input. A new method `StartProcess` has been added to check the validity of the `_processPath` and start the process if valid. Also, the `WaitAndRetryAsync` method in the `AsyncRetryPolicy` has been slightly refactored for better readability.