Adjust `MudItem` size in `ScriptExecutorDialog` and `UpdateDialog`
Adjusted the size of the `MudItem` component in `ScriptExecutorDialog.razor` and `UpdateDialog.razor`. The size was previously set to 4 and 6, but has been changed to 3 and 8 in `ScriptExecutorDialog.razor`, and 3 and 9 in `UpdateDialog.razor`. This change affects the layout of the components on the page, making the `MudItem` component smaller in the first case and larger in the second case. No functionality changes have been made, only layout adjustments.