Updated MapNetworkDrive method and its usage
Updated the method signature of `MapNetworkDrive` in `INetworkDriveService` to return a boolean value indicating the success or failure of the operation. Adjusted the usage of `MapNetworkDrive` in `HostUpdater` to check the return value and log an error message if the operation fails. Also, modified the implementation of `MapNetworkDrive` in `NetworkDriveService` to return a boolean value based on the success or failure of the mapping operation, replacing the previous exception throwing mechanism.