Updated log message in HostUpdater class
Updated the log message in the `HostUpdater` class to provide more informative feedback to the user. Previously, the message would only inform the user that downgrades were not allowed if the current version was up to date or newer than the update version. The new log message now also instructs the user on how to allow downgrades by using the --allow-downgrade option. This change aims to provide more helpful feedback to users wishing to downgrade their version.